Fix compiler warnings reported by Sebastian Reitenbach <sebastia@l00-bugdead-prods.de>
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36165 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSDragView.m
@@ -497,7 +497,7 @@ static	GSDragView *sharedDragView = nil;
 
   NSDebugLLog (@"NSDragging",
                @"drag, operation, target mask = (%x, %x, %x), dnd aware = %d\n",
-               dragMask, operationMask, targetMask,
+               (unsigned int)dragMask, (unsigned int)operationMask, (unsigned int)targetMask,
                (targetWindowRef != 0));
   
   if (cursors == nil)
